+++ include/poisson.h
@@ -1,0 +1,101 @@
+#ifndef POISSON_H
+#define POISSON_H
+
+#include <deal.II/base/function.h>
+#include <deal.II/base/function_lib.h>
+#include <deal.II/base/function_parser.h>
+#include <deal.II/base/logstream.h>
+#include <deal.II/base/parameter_handler.h>
+#include <deal.II/base/quadrature_lib.h>
+#include <deal.II/base/parsed_convergence_table.h>
+
+#include <deal.II/dofs/dof_handler.h>
+#include <deal.II/dofs/dof_tools.h>
+
+#include <deal.II/fe/fe_q.h>
+#include <deal.II/fe/fe_values.h>
+
+#include <deal.II/grid/grid_generator.h>
+#include <deal.II/grid/tria.h>
+#include <deal.II/grid/grid_refinement.h>
+
+#include <deal.II/numerics/data_out.h>
+#include <deal.II/numerics/matrix_tools.h>
+#include <deal.II/numerics/vector_tools.h>
+
+#include <deal.II/lac/affine_constraints.h>
+#include <deal.II/lac/dynamic_sparsity_pattern.h>
+#include <deal.II/lac/full_matrix.h>
+#include <deal.II/lac/precondition.h>
+#include <deal.II/lac/solver_cg.h>
+#include <deal.II/lac/sparse_matrix.h>
+#include <deal.II/lac/vector.h>
+
+#include <string>
+#include <fstream>
+#include <iostream>
+
+using namespace dealii;
+
+template <int dim>
+struct PoissonParameters
+{
+  PoissonParameters(const std::string &filename = "poisson.prm");
+  
+  unsigned int fe_degree                 = 1;
+  unsigned int initial_refinement        = 3;
+  unsigned int n_cycles                  = 1;
+  std::string  exact_solution_expression = "cos(pi*x)*cos(pi*y)";
+  std::string  rhs_expression            = "2*pi*pi*cos(pi*x)*cos(pi*y)";
+  std::string  grid_name                 = "hyper_cube";
+  std::string  grid_arguments            = "0 : 1 : false";
+  std::string  output_file_name          = "poisson_solution";
+  std::string  output_directory          = ".";
+
+  FunctionParser<dim> exact_solution;
+  FunctionParser<dim> rhs_function;
+
+  mutable ParsedConvergenceTable convergence_table;
+
+  ParameterHandler prm;
+
+  double ref_frac = 0.25; 
+  double coarse_frac = 0.25;
+};
+
+template <int dim>
+class Poisson
+{
+public:
+  Poisson(const PoissonParameters<dim> &par);
+  void
+  run();
+
+private:
+  void
+  make_grid();
+  void
+  setup_system();
+  void
+  assemble_system();
+  void
+  solve();
+  void
+  output_results(const unsigned int cycle) const;
+
+  const PoissonParameters<dim> &par;
+
+  Triangulation<dim> triangulation;
+  FE_Q<dim>          fe;
+  DoFHandler<dim>    dof_handler;
+
+  AffineConstraints<double> constraints;
+
+  SparsityPattern      sparsity_pattern;
+  SparseMatrix<double> system_matrix;
+
+  Vector<double> solution;
+  Vector<double> system_rhs;
+};
+
+#endif //POISSON_H
